backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
garminconnect
pydantic
pytest
pytest-asyncio
